db_Mongo/requirements.txt
Faker==37.4.2
pymongo==4.13.2
python-dotenv==1.1.1
